deployment/hf_spaces/requirements_hf.txt
torch
torchvision
gradio
opencv-python-headless
scikit-image
numpy
Pillow
